requirements.txt
python
python-dotenv
streamlit
google-genai 